Rework dialogs
Dialog components like `NewLogbookDialog` now must be contained
in `Dialog` and `Dialog.Content` giving control of whether
or not the dialog is open to the parent. However, we still
want the dialog component to control the look of the dialog which
is why `Dialog.Content` was split into `Dialog.Content` which
only renders its children when the dialog is open and `Dialog.Window`
which controls the look of the dialog.