Fix image previews and attachment downloads in development
In prod, the access token cookie will be transformed into a header for
API server to consume. However, in development we have to manually set
the header, which we did through our custom fetch function. However,
this doesn't work with images and downloads which don't not use the
fetch function and rely on the browser. So to get around this, we set a
cookie `dev-slac-vouch` and in the reverse proxy used during
development, transform that cookie to into correct header :tada:.